added _POSIX_C_SOURCE definition for use of clock_gettime(). replaced function pointers in engine.h with actual functions whose implementations change based on platform of compilation. removed engine_glx_start(), engine_glx_update(), and engine_glx_stop(); replaced with engine_start(), engine_update(), and engine_stop(). moved input key definitions to engine.h. freed XVisualInfo(). fixed value out of range runtime error in mathf_noiseN() functions. added delta time implementation for x11

+++ src/engine_mathf.h
@@ -148,20 +148,19 @@ static inline float mathf_angle_delta(const float a, const float b) {
 static inline float mathf_fraction(float x) { return x - mathf_floor(x); }
 
 // Single dimensional pseudo-random noise
-static inline float mathf_noise1(unsigned int x) {
+static inline float mathf_noise1(float x) {
   float wave = mathf_sin(x * 53) * 6151;
   return mathf_fraction(wave);
 }
 
 // Two dimensional pseudo-random noise
-static inline float mathf_noise2(unsigned int x, unsigned int y) {
+static inline float mathf_noise2(float x, float y) {
   float wave = mathf_sin(x * 53 + y * 97) * 6151;
   return mathf_fraction(wave);
 }
 
 // Three dimensional pseudo-random noise
-static inline float mathf_noise3(unsigned int x, unsigned int y,
-                                 unsigned int z) {
+static inline float mathf_noise3(float x, float y, float z) {
   float wave = mathf_sin(x * 53 + y * 97 + z * 193) * 6151;
   return mathf_fraction(wave);
 }
